requirements.txt
pandas
psycopg2-binary
apache-airflow-providers-postgres
apache-airflow-providers-amazon
python-dotenv
presidio_analyzer
presidio_anonymizer
openai
